+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.3.0
 
 - Fix two-tap bug on verse grid chooser.
+- Shortcut for navigating to verse 1 in the grid verse chooser is to tap the chapter title.
 
 # 2.2.0 - 2026.09.19